Burger Api. rename and fix search by path
