+++ internal/app/service_provider.go
@@ -5,8 +5,8 @@ import (
 	"database/sql"
 	"fmt"
 	"github.com/RomanLevBy/BurgersAPI/internal/config"
+	getAllBurger "github.com/RomanLevBy/BurgersAPI/internal/http-server/handlers/burger/all"
 	getBurger "github.com/RomanLevBy/BurgersAPI/internal/http-server/handlers/burger/get"
-	getAllBurger "github.com/RomanLevBy/BurgersAPI/internal/http-server/handlers/burger/getAll"
 	saveBurger "github.com/RomanLevBy/BurgersAPI/internal/http-server/handlers/burger/save"
 	getCategory "github.com/RomanLevBy/BurgersAPI/internal/http-server/handlers/category/get"
 	burgerRepo "github.com/RomanLevBy/BurgersAPI/internal/repository/burger/postgres"
